+++ CHANGELOG.md
@@ -4,11 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## [1.0.0] - 2024-10-25 (expected)
+## [1.0.0] - 2024-10-25
 First stable GA release
 
 ### Added
 - `Voice.downloadRecording(String, Path)` method
+- `InputAction.Builder.type` Boolean overload method (preferable to the Collection method)
 
 ### Changed
 - `Sms.wasSuccessfullySent()` now an extension function rather than being part of the client